% Auto-generated: do not edit by hand
\name{board}

\alias{board}

\title{Board component}

\description{

}

\usage{
board(id=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

