% Auto-generated: do not edit by hand
\name{bulmaButton}

\alias{bulmaButton}

\title{BulmaButton component}

\description{

}

\usage{
bulmaButton(id=NULL, value=NULL, size=NULL, color=NULL, rounded=NULL,
loading=NULL, outlined=NULL, inverted=NULL, disabled=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The text displayed in the button.}

\item{size}{A value equal to: 'small', 'medium', 'large'. The size of the button.}

\item{color}{Character. The color of the button.}

\item{rounded}{Logical. Determines whether the button has
rounded corners.}

\item{loading}{Logical. Determines whether the button shows
the "loading" icon.}

\item{outlined}{Logical. Determines whether the button is
outlined.}

\item{inverted}{Logical. Determines whether the button has
inverted colors.}

\item{disabled}{Logical. Determines whether the button is
disabled.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

