% Auto-generated: do not edit by hand
\name{clock}

\alias{clock}

\title{Clock component}

\description{

}

\usage{
clock(id=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

