% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-test.R
\name{.test_javascript}
\alias{.test_javascript}
\title{test javascript}
\usage{
.test_javascript(...)
}
\arguments{
\item{...}{Any other argument.}
}
\description{
test javascript component.
}
\keyword{internal}
