% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-test.R
\name{.test_server}
\alias{.test_server}
\title{test server}
\usage{
.test_server(input, output, session, ...)
}
\arguments{
\item{input, output, session}{Arguments passed from mdoule's server.}

\item{...}{Any other argument.}
}
\description{
test server component.
}
\keyword{internal}
