% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-test.R
\name{.test_ui}
\alias{.test_ui}
\title{test ui}
\usage{
.test_ui(ns, ...)
}
\arguments{
\item{ns}{Shiny's namespace function.}

\item{...}{Any other argument.}
}
\description{
test UI component.
}
\keyword{internal}
