% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{namespace}
\alias{namespace}
\title{Namespace}
\usage{
namespace(strings, ns, env = parent.frame())
}
\arguments{
\item{strings}{Vector of string to namespace.}

\item{ns}{Namespace function.}

\item{...}{Environment to evaluate string.}
}
\description{
Namespace a string.
}
\section{Transformers}{

\itemize{
\item \code{class}: transforms \verb{\{\{ class red \}\}} into \code{.namespace-red}
\item \code{id}: transforms \verb{\{\{ id red \}\}} into \verb{#namespace-red}
\item \code{ns}: transforms \verb{\{\{ ns red \}\}} into \code{namespace-red}
\item \code{json}: transforms \verb{\{\{ json letters[1:2] \}\}} into \verb{['a', 'b']}
}
}

