% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_component}
\alias{remove_component}
\title{Remove component}
\usage{
remove_component(name, check = interactive())
}
\arguments{
\item{name}{Name of component to detele.}

\item{check}{Whether to run an interactive check.}
}
\description{
Remove a component from the package.
}
