% Auto-generated: do not edit by hand
\name{square}

\alias{square}

\title{Square component}

\description{

}

\usage{
square(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

