% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-generated-test.R
\name{test_server}
\alias{test_server}
\title{Test generated server}
\usage{
test_server(id, ...)
}
\arguments{
\item{id}{ID of module}

\item{...}{Any arguments to pass to .test_server()}
}
\description{
Server generated from: .test_server()
}
\keyword{internal}
