% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-generated-test.R
\name{test_ui}
\alias{test_ui}
\title{Test generated UI}
\usage{
test_ui(id, ...)
}
\arguments{
\item{id}{ID of module}

\item{...}{Any arguments to pass to .test_ui()}
}
\description{
UI generated from: .test_ui()
}
\keyword{internal}
