% Auto-generated: do not edit by hand
\name{toggle}

\alias{toggle}

\title{Toggle component}

\description{

}

\usage{
toggle(id=NULL, label=NULL, value=NULL, color=NULL, rounded=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}

\item{color}{Character. }

\item{rounded}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

