% Auto-generated: do not edit by hand
\name{bulmaBox}

\alias{bulmaBox}

\title{BulmaBox component}

\description{

}

\usage{
bulmaBox(children=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Dash-assigned callback that should be called to report property changes
to Dash, to make them available for callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

